Opprydding og mer testing
