+++ src/test/kotlin/no/nav/dagpenger/regel/periode/PeriodeSpesifikasjonTest.kt
@@ -13,9 +13,74 @@ import no.nav.dagpenger.events.inntekt.v1.KlassifisertInntekt
 import no.nav.dagpenger.events.inntekt.v1.KlassifisertInntektMåned
 import no.nav.nare.core.evaluations.Evaluering
 import no.nav.nare.core.evaluations.Resultat
+import no.nav.nare.core.specifications.Spesifikasjon
 import org.junit.jupiter.api.Test
 
 internal class PeriodeSpesifikasjonTest {
+
+    fun identifikatorer(spesifikasjoner: List<Spesifikasjon<Fakta>>): Set<String> = (spesifikasjoner.map { it.identifikator }.toSet() + spesifikasjoner.flatMap { identifikatorer(it.children) }).toSet()
+
+    @Test
+    fun `Ordinær består av ordinær`() {
+
+        val expectedIdentifikatorer = setOf<String>(
+            "ORDINÆR_12_52",
+            "ORDINÆR_36_52",
+            "ORDINÆR_12_52_FANGSTOGFISK",
+            "ORDINÆR_36_52_FANGSTOGFISK",
+            "ORDINÆR_12_104",
+            "ORDINÆR_36_104",
+            "ORDINÆR_12_104_FANGSTOGFISK",
+            "ORDINÆR_36_104_FANGSTOGFISK"
+        )
+        val identifikatorer = identifikatorer(ordinær.children)
+        expectedIdentifikatorer shouldBe identifikatorer
+    }
+
+    @Test
+    fun `Særregel består av verneplikt eller lærling`() {
+
+        val expectedIdentifikatorer = setOf<String>(
+            "VERNEPLIKT",
+            "LÆRLING"
+        )
+        val identifikatorer = identifikatorer(særregel.children)
+        expectedIdentifikatorer shouldBe identifikatorer
+    }
+
+    @Test
+    fun `Skal returnere 52 som antall uker`() {
+        val evaluering = Evaluering.ja("52")
+        mapEvalureringResultatToInt(evaluering).max() shouldBe 52
+    }
+
+    @Test
+    fun `Skal returnere evalueringene som int`() {
+        val evaluering = listOf(Evaluering.ja("52"), Evaluering.ja("26"))
+
+        val resultat = evaluering.flatMap { mapEvalureringResultatToInt(it) }
+
+        listOf(52, 26) shouldBe resultat
+    }
+
+    @Test
+    fun `Skal returnere høyeste periode som har evaluering lik JA`() {
+
+        val evaluering = listOf(Evaluering.ja("52"), Evaluering.ja("26"), Evaluering.nei("104"))
+
+        val rotEvaluering = Evaluering(Resultat.JA, "52", children = evaluering)
+        val fakta = Fakta(
+            Inntekt("123", emptyList(), sisteAvsluttendeKalenderMåned = YearMonth.of(2019, 4)),
+            bruktInntektsPeriode = null,
+            verneplikt = false,
+            fangstOgFisk = false,
+            beregningsDato = LocalDate.of(2019, 5, 20),
+            grunnlagBeregningsregel = "bla",
+            lærling = false
+        )
+
+        finnHøyestePeriodeFraEvaluering(rotEvaluering) shouldBe 52
+    }
 
     @Test
     fun ` Ordinær skal ikke behandle lærling under koronatid`() {
